Completed duplicate checking
It's a shitty way of doing it, but this will find all duplicates and
write them to a separate txt file. I couldn't for the life of me figure
out how to remove all but one duplicate, but this is an easy (shitty)
workaround. Have to manually add one of the duplicates back if you so
desire.