Fixed bug in GetMostUsedChamp, updated ignore
Gitignore now ignores all files added to the DEPRECATED folder

A bug in finding the most used champion occurred because I overlooked
the fact that the Get Ranked Stats API call returns a summoner with ID 0
at the end of the stats, which is cumulative data of all the summoners.
This ID 0 caused a serious bug that would exclude dozens of summoners.